requirements.txt
pymongo~=4.13.2
pandas~=2.3.1
numpy~=1.26.2
joblib~=1.5.1
wandb~=0.21.0
optuna~=4.4.0
scikit-learn~=1.7.0
transformers~=4.53.2
chempy~=0.9.0
tqdm~=4.67.1
requests~=2.32.4
pubchempy~=1.0.4
rdkit~=2025.3.3
pymatgen~=2025.6.14
jsonlines~=4.0.0
pyyaml~=6.0.2
setuptools~=80.9.0
langchain~=0.3.26
pydantic~=2.11.7
fuzzywuzzy~=0.18.0
opencv-python~=4.6.0.66
doc2docx~=0.2.4
docx~=0.2.4
markdownify~=1.1.0
pdf2image~=1.17.0
openai~=1.95.1
spacy~=3.8.7
datasets~=4.0.0
beautifulsoup4~=4.13.4
tokenizers~=0.21.2
python-docx==1.1.0
opencv-python==4.6.0.66
paddleocr==2.7.0.3
weaviate-client==4.4b0
paddlepaddle-gpu=2.3.0
importlib==1.0.4
torch